requirements.txt
pydantic>=1.10
python-dotenv>=1.0
tenacity>=8.2.0
rapidfuzz>=2.14.0
requests>=2.28.0
groq>=0.1.0